+++ database/migrations/20260628144636_add_ssrfwhitelist_to_configuration_table.php
@@ -1,0 +1,21 @@
+<?php
+
+use Movim\Migration;
+use Illuminate\Database\Schema\Blueprint;
+
+class AddSsrfwhitelistToConfigurationTable extends Migration
+{
+    public function up()
+    {
+        $this->schema->table('configuration', function (Blueprint $table) {
+            $table->text('ssrfwhitelist')->nullable();
+        });
+    }
+
+    public function down()
+    {
+        $this->schema->table('configuration', function (Blueprint $table) {
+            $table->dropColumn('ssrfwhitelist');
+        });
+    }
+}